requirements.txt
asgiref==3.5.2
Django==4.1.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
